refactor: re-arrange folder for schema
